Rectify cpufreq module to map cpu frequencies to corresponding cores on powerpc
The cpus under sysfs reflect all cpu threads on the system.

We use the fact that cpu-cores under /proc/device-tree/cpus/ have a reg
property which is same as the /topology/core-id property on all its
threads and that only the online thread has /online property in sysfs
cpus as 1, to map the (first) online sysfs cpu thread with device-tree
cpu core and have the cpu frequencies pulled from sysfs thread to the cpu.

Signed-off-by: Chandni Verma <chandni@linux.vnet.ibm.com>