AArch64: cpuinfo: Remove redundant cpu caps loop
CPU capabilities should be added only to the current CPU core,
and not recursively to all previous CPU cores.

For a 48 core node, this parsing added up to 30 seconds (8 lines
per core in /proc/cpuinfo resulted in 384 lines to parse).
For a 96 core node, parsing takes a few minutes.

After this change, parsing for the above systems takes less than
1 second.

Closes: https://bugs.launchpad.net/ubuntu/+source/lshw/+bug/1582181

Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>