Fix use of LONG_BIT
LONG_BIT is not a sysconf value, it is either 32 or 64. Using it as
a sysconf value will give weird results.

Also sysconf(_SC_LONG_BIT) is useless: it will always return a value
equal to LONG_BIT: it's either compiled 32-bit or 64-bit so a runtime
lookup doesn't make sense. For this reason, musl has removed the
definition of _SC_LONG_BIT.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>