Fix abnormal termination from throwing std::bad_alloc().
On my system, recent updates made me start getting this exception when I
forget and run lshw as a normal user.  Investigating, I see that it is
because /sys/firmware/dmi/tables/smbios_entry_point is readable only as
root.  This means that tellg() returns -1, and the vector attempts to
allocate a LOT of memory.

I wasn't sure if the second if( !ep_stream) was still necessary (after
the call the read() ), so I left it.