Fix const correctness errors introduced in d3c66a6b2.
The dirname and basename functions from libgen.h may (or not) change the
string passed to them, which is why their parameter is "char *", not
"const char *". Using "const_cast<char *>" so the compiler didn't
complain was the wrong change to make, since the underlying data of
those strings would be changed anyway.

This commit fixes that by adding a basename() wrapper in osutils.cc,
based on the existing dirname() wrapper, and using it where applicable.