Fix getting size of memory banks <32GiB
Due to a regression introduced by
8ff1efbc080c804f65f2557f9de3a5057982173d, no size was recorded for
memory banks <32GiB in size on systems with an SMBIOS version of 2.7 or
later. On these systems the Type 17 size field from SMBIOS was only
recorded if the extended size field was used.

Modify the code to use the regular size field whenever it is valid (not
0xFFFF) and not set to 0x7FFF on versions >2.7 (indicating the extended
size field is in use).