Only way I could find to notify you that your issue tracker is broken
CAPTCHA error trying to submit bug reports on your website at https://ezix.org/project/newticket#ticket

I have a couple of bugs I've been sat on for years because I haven't been able to get the CAPTCHA to work on your website. This is the only way I could think of to notify you of the problem.