Set correct logical name of NVMe disks with native multipath enabled
When native NVMe multipath is enabled for NVMe devices, logical name for
NVMe disks is displayed as follows:

$ lshw -c disk | grep "logical name"
       logical name: nvme1c1n1
       logical name: nvme2c2n1
       logical name: nvme3c3n1
       logical name: /dev/sda
       logical name: /dev/sdb

This is happening because, when native multipath is enabled, NVMe device's
sysfs directory path has subdirectories like nvme1c1n1, nvme2c2n1, etc..
If that subdirectory is present in /dev, then code adds /dev to the
subdirectory name and prints it as logical name. Else, it prints the
subdirectory name itself as logical name. Since nvme1c1n1 directory is
present in sysfs directory, and since /dev does not have nvme1c1n1,
code is currently printing logical name as nvme1c1n1 itself.

nvme list command is listing the device node under /dev as follows:

$ nvme list | awk '{print $1}'
Node
---------------------
/dev/nvme3n1
/dev/nvme2n1
/dev/nvme1n1

This patch determines the correct logical name of NVMe disks when native
multipath is enabled, by removing the controller name from the namespace
name and concatenating the resulting string with /dev/. If the resulting
string exists, then set it as the logical name of the NVMe disk.

With this patch, logical names are displayed as follows:

$ ./src/lshw -c disk | grep "logical name"
       logical name: /dev/nvme1n1
       logical name: /dev/nvme2n1
       logical name: /dev/nvme3n1
       logical name: /dev/sda
       logical name: /dev/sdb

Reported-by: Naresh Bannoth <nbannoth@in.ibm.com>
Suggested-by: Wen Xiong <wenxiong@linux.ibm.com>
Signed-off-by: Sathvika Vasireddy <sv@linux.ibm.com>