Fix SAST uninitialized implementation
There is no clear constructor of statement_i, as a result, the
implementation->db is uninitialized, which leaves a potential fail of
implementation->db->implementation referencing.

By checking the code context, the implementation is likely already been
initialized, and unlikely the "new statement_i" to be called. So just
remove the relating code.

Signed-off-by: Tao Liu <ltao@redhat.com>